Drop redundant int() around math.ceil
math.ceil returns int in py3; the cast is a py2 leftover. Pyrefly flags it
as unnecessary-type-conversion.

+++ src/picosvg/arc_to_cubic.py
@@ -139,7 +139,7 @@ def _arc_to_cubic(arc: EllipticalArc) -> Iterator[Tuple[Point, Point, Point]]:
     # Some results of atan2 on some platform implementations are not exact
     # enough. So that we get more cubic curves than expected here. Adding 0.001f
     # reduces the count of sgements to the correct count.
-    num_segments = int(ceil(fabs(arc_params.theta_arc / (PI_OVER_TWO + 0.001))))
+    num_segments = ceil(fabs(arc_params.theta_arc / (PI_OVER_TWO + 0.001)))
     for i in range(num_segments):
         start_theta = arc_params.theta1 + i * arc_params.theta_arc / num_segments
         end_theta = arc_params.theta1 + (i + 1) * arc_params.theta_arc / num_segments
